Fix unused code warnings.
